Move guard to…well, guards file
